Feature/input validation

+++ src/main/java/com/mishas/stuff/mta/service/impl/AccountService.java
@@ -7,7 +7,6 @@ import com.mishas.stuff.mta.persistence.model.Transfer;
 import com.mishas.stuff.mta.service.IAccountService;
 import com.mishas.stuff.mta.persistence.model.TransferResult;
 import com.mishas.stuff.mta.web.dto.AccountDto;
-import com.mishas.stuff.mta.web.dto.TransferDto;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -20,7 +19,10 @@ public class AccountService implements IAccountService {
     @Override
     public AccountDto get(long id) {
         LOGGER.info("getting resource with ID: " + id);
-        return new AccountDto(accountRepository.get(id));
+        if (accountExists(id)){
+            return new AccountDto(accountRepository.get(id));
+        }
+        return new AccountDto();
     }
 
     @Override
@@ -32,14 +34,20 @@ public class AccountService implements IAccountService {
     @Override
     public AccountDto update(long id, AccountDto resource) {
         LOGGER.info("updating a resource with Id: " + id);
-        Account entity = accountRepository.update(id, new Account(resource));;
-        return new AccountDto(entity);
+        if (accountExists(id)){
+            Account entity = accountRepository.update(id, new Account(resource));;
+            return new AccountDto(entity);
+        }
+        return new AccountDto();
+
     }
 
     @Override
     public void delete(long id) {
         LOGGER.info("deleting a resource with Id: " + id);
-        accountRepository.delete(id);
+        if (accountExists(id)){
+            accountRepository.delete(id);
+        }
     }
 
     public boolean accountExists(long id) {
